+++ MyStore.Application/Request/UpdateCartItemRequest.cs
@@ -2,7 +2,7 @@
 {
     public class UpdateCartItemRequest
     {
-        public int? SizeId { get; set; }
+        public long? SizeId { get; set; }
         public int? Quantity { get; set; }
     }
 }
