+++ MyStore.Domain/Entities/Image.cs
@@ -2,9 +2,9 @@
 {
     public class Image : IBaseEntity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public Product Product { get; set; }
 
         public string ImageUrl { get; set; }
